move split_run_length_lists to flipper conversion file
